requirements.txt
# Backend Agentes Peers - Dependências Python (Windows Compatible)
# Execute: pip install -r requirements.txt

# === Core Framework (versões compatíveis com Windows) ===
fastapi>=0.100.0,<1.0.0
uvicorn[standard]>=0.20.0,<1.0.0
pydantic>=2.0.0,<3.0.0
python-multipart>=0.0.5,<1.0.0

# === AI & APIs ===
openai>=1.10.0,<2.0.0
PyGithub>=1.59.0,<3.0.0

# === Utilities ===
python-dotenv>=1.0.0,<2.0.0
requests>=2.25.0,<3.0.0

# === Essential Dependencies ===
httpx>=0.24.0,<1.0.0
h11>=0.14.0,<1.0.0
starlette>=0.27.0,<1.0.0
anyio>=3.7.0,<5.0.0

# === Type Support ===
typing-extensions>=4.5.0

# === Legacy Support ===
Deprecated>=1.2.0,<2.0.0

# === Windows Specific Notes ===
# Se ainda houver problemas, tente:
# pip install --upgrade pip setuptools wheel
# pip install --only-binary=:all: -r requirements.txt